requirements.txt
amqp==5.1.1
asgiref==3.6.0
attrs==22.2.0
billiard==3.6.4.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
crispy-bootstrap5==0.7
Django==4.1.7
django-cors-headers==3.14.0
django-crispy-forms==2.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
exceptiongroup==1.1.1
Faker==12.0.1
idna==3.4
iniconfig==2.0.0
kombu==5.2.4
mixer==7.2.2
packaging==23.0
pluggy==1.0.0
prompt-toolkit==3.0.38
pytest==7.2.2
python-dateutil==2.8.2
python-status==1.0.1
pytz==2022.7.1
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.7
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
